% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_peakPeriod.R
\name{add_peakPeriod}
\alias{add_peakPeriod}
\title{Codes morning/evening peak periods using \code{rDateTimeUTC} in the data.table}
\usage{
add_peakPeriod(
  dt,
  dateTime = rDateTimeUTC,
  t1 = "07:00",
  t2 = "09:00",
  t3 = "16:00",
  t4 = "20:00"
)
}
\arguments{
\item{dt}{a data table with a summary variable you want to average}

\item{dateTime}{the variable which is a nice R dateTime for passing to hms::as_hms()}

\item{t1}{start of morning peak in 24 hour clock e.g. "07:00" (default)}

\item{t2}{end of morning peak in 24 hour clock e.g. "08:30" (default)}

\item{t3}{start of morning peak in 24 hour clock e.g. "16:00" (default)}

\item{t4}{start of morning peak in 24 hour clock e.g. "20:00" (default)}
}
\description{
\code{add_peakPeriod} adds a factor column \code{dv_peakPeriod} to the input data.table
}
\seealso{
Other utils: 
\code{\link{add_season}()},
\code{\link{gzip_file}()},
\code{\link{load_libraries}()},
\code{\link{make_flexTable}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils}
