% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_season.R
\name{add_season}
\alias{add_season}
\title{Add season to a data.table depending on hemisphere}
\usage{
add_season(dt, dateVar, h)
}
\arguments{
\item{dt}{the data table}

\item{dateVar}{the column in the dt which is a date that can be passed to lubridate::month()}

\item{h}{hemisphere: North (N) or South (S)?}
}
\description{
\code{add_season()} returns a dt with hemisphere season (Winter, Spring, Summer, Autumn)
added - assume temperature latitudes
}
\seealso{
Other utils: 
\code{\link{add_peakPeriod}()},
\code{\link{gzip_file}()},
\code{\link{load_libraries}()},
\code{\link{make_flexTable}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils}
