% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDSTBreakDate.R
\name{addDSTBreakDate}
\alias{addDSTBreakDate}
\title{\code{addDSTBreakDate} adds label for DST break 31/3/2020 where x axis is a date. Should work in any time-zone}
\usage{
addDSTBreakDate(p, yMin, yMax)
}
\arguments{
\item{p}{the plot to add them to}

\item{yMin}{the smallest y value}

\item{yMax}{the largest y value}
}
\description{
\code{addDSTBreakDate} adds label for DST break 31/3/2020 where x axis is a date. Should work in any time-zone
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
