% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDSTBreakDateTime.R
\name{addDSTBreakDateTime}
\alias{addDSTBreakDateTime}
\title{\code{addDSTBreakDateTime} adds label for DST break 31/3/2020 where x axis is a dateTime. Should work in any time-zone}
\usage{
addDSTBreakDateTime(p, yMin, yMax)
}
\arguments{
\item{p}{the plot to add them to}

\item{yMin}{the smallest y value}

\item{yMax}{the largest y value}
}
\description{
\code{addDSTBreakDateTime} adds label for DST break 31/3/2020 where x axis is a dateTime. Should work in any time-zone
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
