% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNZSeason.R
\name{addNZSeason}
\alias{addNZSeason}
\title{Add NZ season to a data.table}
\usage{
addNZSeason(dt, date)
}
\arguments{
\item{dt}{the data table}

\item{date}{the column in the dt which is a date that lubridate::month() will work on}
}
\description{
\code{addNZSeason} returns a dt with SOUTHERN hemisphere season added.
}
\seealso{
Other utils: 
\code{\link{addLockdownRect}()},
\code{\link{addSeason}()},
\code{\link{alignDates}()},
\code{\link{getDuration}()},
\code{\link{gzipFile}()},
\code{\link{loadLibraries}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{setPeakPeriod}()},
\code{\link{setup}()},
\code{\link{tidyNum}()}

Other NZ: 
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{makeNZYearlyData}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{NZ}
\concept{utils}
