% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSeason.R
\name{addSeason}
\alias{addSeason}
\title{Add season to a data.table depending on hemisphere}
\usage{
addSeason(dt, dateVar, h)
}
\arguments{
\item{dt}{the data table}

\item{dateVar}{the column in the dt which is a date that lubridate::month() will work on}

\item{h}{hemisphere: North (N) or South (S)?}
}
\description{
\code{addSeason} returns a dt with hemisphere season (Winter, Spring, Summer, Autumn) added - assume temperature latitudes
}
\seealso{
Other utils: 
\code{\link{addLockdownRect}()},
\code{\link{addNZSeason}()},
\code{\link{alignDates}()},
\code{\link{getDuration}()},
\code{\link{gzipFile}()},
\code{\link{loadLibraries}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{setPeakPeriod}()},
\code{\link{setup}()},
\code{\link{tidyNum}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils}
