% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVeDayDate.R
\name{addVeDayDate}
\alias{addVeDayDate}
\title{\code{addVeDayDate} adds extended weekend shading and label for VE day 2020 where x axis is a date. Should work in any time-zone}
\usage{
addVeDayDate(p, yMin, yMax)
}
\arguments{
\item{p}{the plot to add them to}

\item{yMin}{the smallest y value}

\item{yMax}{the largest y value}
}
\description{
\code{addVeDayDate} adds extended weekend shading and label for VE day 2020 where x axis is a date. Should work in any time-zone
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
