% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVeDayDateTime.R
\name{addVeDayDateTime}
\alias{addVeDayDateTime}
\title{\code{addVeDayDateTime} adds extended weekend shading and label for VE day 2020 where x axis is a dateTime. 
Should work in any time-zone.}
\usage{
addVeDayDateTime(p, yLoc, yMin, yMax, fill = "grey50", alpha = 0.3)
}
\arguments{
\item{p}{the plot to add them to}

\item{yLoc}{the relative location of the label (% of max plot height)}

\item{yMin}{the smallest y value}

\item{yMax}{the largest y value#'}

\item{fill}{fill colour (grey50)}

\item{alpha}{alpha value (0.3)}
}
\description{
\code{addVeDayDateTime} adds extended weekend shading and label for VE day 2020 where x axis is a dateTime. 
Should work in any time-zone.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
