% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWhitsunDateTime.R
\name{addWhitsunDateTime}
\alias{addWhitsunDateTime}
\title{\code{addWhitsunDate} adds extended weekend shading and label for UK Whit Sunday (late spring bank holiday) 
where x axis is a dateTime. Should work in any time-zone}
\usage{
addWhitsunDateTime(p, yLoc, yMin, yMax, fill = "grey50", alpha = 0.5)
}
\arguments{
\item{p}{the plot to add them to}

\item{yLoc}{the reltative height of the label (so you can stagger them on a plot)}

\item{yMin}{the smallest y value}

\item{yMax}{the largest y value}

\item{fill}{fill colour (grey50)}

\item{alpha}{alpha value (0.5)}
}
\description{
\code{addWhitsunDate} adds extended weekend shading and label for UK Whit Sunday (late spring bank holiday) 
where x axis is a dateTime. Should work in any time-zone
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
