% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignDates.R
\name{alignDates}
\alias{alignDates}
\title{Re-align 2017-2019 dates so their weekdays match to 2020}
\usage{
alignDates(dt)
}
\arguments{
\item{dt}{the data.table to align}
}
\description{
\code{alignDates} returns a dt with `fixedDate` set to the date needed in 2017-2019 to match the nearest weekday
in 2020. Use this for plots which try to compare daily values over time.
}
\details{
The data.table to align *must* have `rDateTimeUTC` correctly set.
}
\seealso{
Other data: 
\code{\link{loadUKESOYearlyGenData}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{data}
