% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNZEmbEA.R
\name{cleanNZEmbEA}
\alias{cleanNZEmbEA}
\title{\code{cleanNZEmbEA} cleans up the raw embedded gen file from the NZ EA data website}
\usage{
cleanNZEmbEA(dt)
}
\arguments{
\item{dt}{the data.table to clean up}
}
\description{
These are not in pretty form so we clean them up to a long form file and fix the dateTimes.
}
\details{
NB: rDateTime will be NA for TP49/50 which correspond to the DST breaks.
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanNZGxpEA}()},
\code{\link{cleanUkEmbeddedESO}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getNZGxpEA}()},
\code{\link{getNZGxpMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{getUkGridESO}()},
\code{\link{loadNZEAYearlyEmbeddedGenData}()},
\code{\link{loadNZEAYearlyGridGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{loadUKEmbeddedGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}

Other embedded: 
\code{\link{cleanUkEmbeddedESO}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{makeNZYearlyData}()},
\code{\link{saveUkEmbeddedESO}()}

Other NZ: 
\code{\link{cleanNZGridEA}()},
\code{\link{cleanNZGxpEA}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getNZGxpEA}()},
\code{\link{getNZGxpMeta}()},
\code{\link{makeNZYearlyData}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{NZ}
\concept{data}
\concept{embedded}
