% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNZGridEA.R
\name{cleanNZGridEA}
\alias{cleanNZGridEA}
\title{\code{cleanNZGridEA} cleans up the raw grid gen file from the NZ EA data website}
\usage{
cleanNZGridEA(dt)
}
\arguments{
\item{dt}{the data.table to clean up}
}
\description{
These are not in pretty form so we clean them up to a long form file and fix the dateTimes forcing to NZT just in case.
}
\details{
Note that rDateTimeNZT will be NA for the DST breaks which equate to TP49/50. We really dislike DST breaks.
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getUkGridESO}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}

Other grid: 
\code{\link{getNZEmbData}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{makeNZYearlyData}()}

Other NZ: 
\code{\link{addNZSeason}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{makeNZYearlyData}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{NZ}
\concept{data}
\concept{grid}
