% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUkEmbeddedESO.R
\name{cleanUkEmbeddedESO}
\alias{cleanUkEmbeddedESO}
\title{\code{cleanUkEmbeddedESO} cleans up UK embedded ESO data adding latest the update to the historical data we
previously manually downloaded. Not ideal. Hopefully NG_ESO will fix the data discontinuity at some point.}
\usage{
cleanUkEmbeddedESO(dt, path)
}
\arguments{
\item{dt}{the data.table to clean (returned by \code{getUkEmbeddedESO})}

\item{rawPath}{ehere to find the pre-downloaded historical data (manually downloaded. Don't ask)}
}
\description{
- adds proper dateTime 
 - keeps just solar & wind power
 
 @import data.table
 @import lubridate
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanNZGxpEA}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getNZGxpEA}()},
\code{\link{getNZGxpMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{getUkGridESO}()},
\code{\link{loadEmbeddedGenData}()},
\code{\link{loadNZEAYearlyEmbeddedGenData}()},
\code{\link{loadNZEAYearlyGridGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}

Other embedded: 
\code{\link{cleanNZEmbEA}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{makeNZYearlyData}()},
\code{\link{saveUkEmbeddedESO}()}

Other uk: 
\code{\link{cleanUkGridESO}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{data}
\concept{embedded}
\concept{uk}
