% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDailyPcComparePlot.R
\name{createDailyPcComparePlot}
\alias{createDailyPcComparePlot}
\title{Creates a daily mean half-hourly % difference plot from the data.table given}
\usage{
createDailyPcComparePlot(dt, yVar, yCap)
}
\arguments{
\item{dt}{the data, assumed to be the aligned data (use alignDates() to create this)}

\item{yVar}{the variable you want to plot}

\item{yCap}{the caption for the y axis}
}
\description{
`createDailyPcComparePlot` returns a plot which calculates the daily mean of yVar and plots the % difference between the value for 2020 and
 2017-2019 to compare the values for the same date over previous years. This assumes you pass in the aligned-date data.
}
\seealso{
Other plot: 
\code{\link{createDailyMeanComparePlot}()},
\code{\link{createRecentDateTimePlot}()},
\code{\link{createRecentHalfHourlyProfilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{plot}
