% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRecentDateTimePlot.R
\name{createRecentDateTimePlot}
\alias{createRecentDateTimePlot}
\title{Creates a half-hourly plot from the data.table given}
\usage{
createRecentDateTimePlot(dt, dateTime = "rDateTimeUTC", yVar, yCap, yDiv = 1)
}
\arguments{
\item{dt}{the data}

\item{dateTime}{the dateTime variable to use (allows for aligned dateTimes etc). Default = `rDateTimeUTC`}

\item{yVar}{the variable you want to plot}

\item{yCap}{the caption for the y axis}

\item{yDiv}{the value you want to divide yVar by to make the y axis more sensible. Default = `1`}
}
\description{
`createRecentDateTimePlot` returns a plot which plots yVar by dateTime and adds lockdown annotations
}
\seealso{
Other plot: 
\code{\link{createDailyMeanComparePlot}()},
\code{\link{createDailyPcComparePlot}()},
\code{\link{createRecentHalfHourlyProfilePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{plot}
