% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRecentHalfHourlyProfilePlot.R
\name{createRecentHalfHourlyProfilePlot}
\alias{createRecentHalfHourlyProfilePlot}
\title{Creates a half-hourly profile plot by dayfrom the data.table given}
\usage{
createRecentHalfHourlyProfilePlot(
  dt,
  dateTime = "rDateTimeUTC",
  yVar,
  yCap,
  yDiv = 1
)
}
\arguments{
\item{dt}{the data}

\item{yVar}{the variable you want to plot}

\item{yCap}{the caption for the y axis}

\item{yDiv}{the value you want to divide yVar by to make the y axis more sensible. Default = `1`}

\item{dateTimeVar}{the dateTime variable to use (allows for aligned dateTimes etc). Default = `rDateTimeUTC`}
}
\description{
`createRecentHalfHourlyProfilePlot` returns a plot which plots yVar by time of day for each day. Assumes rDateTimeUTC exists.
}
\seealso{
Other plot: 
\code{\link{addLockdownRect}()},
\code{\link{createDailyMeanComparePlot}()},
\code{\link{createDailyPcComparePlot}()},
\code{\link{createRecentDateTimePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{plot}
