% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{findParentDirectory}
\alias{findParentDirectory}
\title{Find the path to Parent Directory}
\usage{
findParentDirectory(Parent)
}
\arguments{
\item{Parent}{the name of the directory to which the function should track back.
Should be the root of the GitHub repository}
}
\description{
Equivalent of \code{findParentDirectory}. Is useful for running a project
  across multiple computers where the project is stored in different directories.
}
\author{
Mikey Harper, \email{m.harper@soton.ac.uk}
}
