% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUkGridESO.R
\name{getUkGridESO}
\alias{getUkGridESO}
\title{\code{getUkGridESO} gets the file from the UK Grid ESO data website. These are not in pretty form.
We need to clean them after this.}
\usage{
getUkGridESO(f, update)
}
\arguments{
\item{f}{the file to get (as a url suitable for data.table::fread())}
}
\description{
\code{getUkGridESO} gets the file from the UK Grid ESO data website. These are not in pretty form.
We need to clean them after this.
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanNZGxpEA}()},
\code{\link{cleanUkEmbeddedESO}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getNZGxpEA}()},
\code{\link{getNZGxpMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{loadEmbeddedGenData}()},
\code{\link{loadNZEAYearlyEmbeddedGenData}()},
\code{\link{loadNZEAYearlyGridGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{data}
