% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzip_file.R
\name{gzip_file}
\alias{gzip_file}
\title{gzip a file}
\usage{
gzip_file(f)
}
\arguments{
\item{f}{the file}
}
\description{
\code{gzip_file()} gzips a file
}
\details{
\itemize{
\item usually one we've just written out.
}
}
\seealso{
Other utils: 
\code{\link{add_peakPeriod}()},
\code{\link{add_season}()},
\code{\link{load_libraries}()},
\code{\link{make_flexTable}()}
}
\author{
Michael Harper, \email{m.harper@soton.ac.uk}

Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils}
