% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gzipFile}
\alias{gzipFile}
\title{gzip a file}
\usage{
gzipFile(f)
}
\arguments{
\item{f}{the file}
}
\description{
\code{gzipFile} gzips a file
}
\details{
- usually one we've just written out.
}
\author{
Michael Harper, \email{m.harper@soton.ac.uk}

Ben Anderson, \email{b.anderson@soton.ac.uk}
}
