% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gbGenMix.R
\name{load_gbGenMix}
\alias{load_gbGenMix}
\title{Make a flex table}
\usage{
load_gbGenMix(url, dataPath, olderThan = 14)
}
\arguments{
\item{url}{the url to the NG-ESO data file at https://data.nationalgrideso.com/carbon-intensity1/historic-generation-mix.
This changes from time to time so best passed as a parameter.}

\item{olderThan}{if the most recent date in the local data is older than this,
re-download. Default is 14.}

\item{localFile}{the path the locally held data (if it exists). File name of
\code{df_fuel_ckan.csv} (as per original) is assumed.}
}
\description{
\code{load_gbGenMix()} loads the GB NG-ESO Historic Generation Mix & Carbon Intensity data.
}
\details{
This is regularly updated at https://data.nationalgrideso.com/carbon-intensity1/historic-generation-mix

The function checks if the version of the data we have stored at \code{dataPath}
(if we have it) is olderThan x days. If so it re-downloads it and tries to save
it at \code{dataPath}. It will attempt to create \code{dataPath} if it doesn't exist.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils data}
