% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libraries.R
\name{load_libraries}
\alias{load_libraries}
\title{Installs and loads packages}
\usage{
load_libraries(..., repo = "https://cran.rstudio.com")
}
\arguments{
\item{...}{A list of packages}

\item{repo}{The repository to load functions from. Defaults to "https://cran.rstudio.com"}
}
\description{
\code{load_libraries()} checks whether the package is already installed,
installing those which are not preinstalled. All the libraries are then loaded.
}
\details{
Especially useful when running on virtual machines where package installation
is not persistent (Like UoS sve). It will fail if the packages need to be
installed but there is no internet access.

NB: in R 'require' tries to load a package but throws a warning & moves on if it's not there
whereas 'library' throws an error if it can't load the package. Hence \code{load_libraries()}
https://stackoverflow.com/questions/5595512/what-is-the-difference-between-require-and-library
}
\seealso{
Other utils: 
\code{\link{add_peakPeriod}()},
\code{\link{add_season}()},
\code{\link{gzip_file}()},
\code{\link{make_flexTable}()}
}
\author{
Luke Blunden, \email{lsb@soton.ac.uk} (original)

Michael Harper \email{m.harper@soton.ac.uk} (revised version)
}
\concept{utils}
