% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadLibraries}
\alias{loadLibraries}
\title{Installs and loads packages}
\usage{
loadLibraries(..., repo = "https://cran.rstudio.com")
}
\arguments{
\item{...}{A list of packages}

\item{repo}{The repository to load functions from. Defaults to "https://cran.rstudio.com"}
}
\description{
\code{loadLibraries} checks whether the package is already installed,
  installing those which are not preinstalled. All the libraries are then loaded.
}
\details{
Especially useful when running on virtual machines where package installation
  is not persistent (Like UoS sve). It will fail if the packages need to be
  installed but there is no internet access.

NB: in R 'require' tries to load a package but throws a warning & moves on if it's not there
  whereas 'library' throws an error if it can't load the package. Hence 'loadLibraries'
  https://stackoverflow.com/questions/5595512/what-is-the-difference-between-require-and-library
}
\author{
Luke Blunden, \email{lsb@soton.ac.uk} (original)

Michael Harper \email{m.harper@soton.ac.uk} (revised version)
}
