% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNZEAYearlyEmbeddedGenData.R
\name{loadNZEAYearlyEmbeddedGenData}
\alias{loadNZEAYearlyEmbeddedGenData}
\title{Load the pre-downloaded yearly NZ EA embedded generation files in to a data.table}
\usage{
loadNZEAYearlyEmbeddedGenData(path, fromYear, update)
}
\arguments{
\item{path}{the folder to look in for the data}

\item{fromYear}{the year to start from (needs to be in the data file names - you did name them sensibly, yes?)}
}
\description{
\code{loadNZEAYearlyEmbeddedGenData} returns a dt with a proper rDateTimeNZT added & set to NZ tzone. The 
embedded generation data from XX does not have a fuel type so we cannot calculate carbon intensity. We also
assume that I means inflow (to the grid - so we add this to the grid generation) and that X means outflow (to the LV network) which 
we don't really care about.
}
\details{
Presumably the sum of X is the sum of consumption across the POCs which happen to have embedded generation. Maybe.
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanNZGxpEA}()},
\code{\link{cleanUkEmbeddedESO}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getNZGxpEA}()},
\code{\link{getNZGxpMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{getUkGridESO}()},
\code{\link{loadEmbeddedGenData}()},
\code{\link{loadNZEAYearlyGridGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{data}
