% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNZEAYearlyGridGenData.R
\name{loadNZEAYearlyGridGenData}
\alias{loadNZEAYearlyGridGenData}
\title{Load the pre-downloaded yearly NZ EA grid generation files in to a data.table}
\usage{
loadNZEAYearlyGridGenData(path, fromYear, update)
}
\arguments{
\item{path}{the folder to look in for the data}

\item{fromYear}{the year to start from (needs to be in the data file names - you did name them sensibly, yes?)}
}
\description{
\code{loadNZEAYearlyGridGenData} returns a dt with a proper rDateTimeNZT added & set to NZ tzone.
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanUkEmbeddedESO}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{getUkGridESO}()},
\code{\link{loadEmbeddedGenData}()},
\code{\link{loadNZEAYearlyEmbeddedGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{data}
