% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flexTable.R
\name{make_flexTable}
\alias{make_flexTable}
\title{Make a flex table}
\usage{
make_flexTable(df, caption = "caption", digits = 1)
}
\arguments{
\item{df}{the data frame or data.table to make into a table}

\item{digits}{the number of digits to include on numeric values (aplies to all numeric colums)}

\item{cap}{a caption}
}
\description{
\code{make_flexTable()} builds a flextable using our preferred defaults.
It returns the flextable so you can then change anything.
}
\seealso{
Other utils: 
\code{\link{add_peakPeriod}()},
\code{\link{add_season}()},
\code{\link{gzip_file}()},
\code{\link{load_libraries}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils}
