% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukGridESO.R
\name{makeUkGridESOYearlyData}
\alias{makeUkGridESOYearlyData}
\title{\code{makeUkGridESOYearlyData} saves the UK Grid ESO data as yearly files to the location you tell it}
\usage{
makeUkGridESOYearlyData(dt, path)
}
\arguments{
\item{dt}{the data.table to save (returned by \code{cleanUkGridESO})}

\item{path}{the folder to put the yearly files in}
}
\description{
\code{makeUkGridESOYearlyData} saves the UK Grid ESO data as yearly files to the location you tell it
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
