% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWeekdayPlot.R
\name{makeWeekdayPlot}
\alias{makeWeekdayPlot}
\title{Creates a plot by weekday from the data.table given}
\usage{
makeWeekdayPlot(dt, yVar, yLab, yDiv)
}
\arguments{
\item{dt}{the data}

\item{yVar}{the variable you want to plot}

\item{yLab}{the label for the y axis}

\item{yDiv}{the value you want to divide yVar by to make the y axis more sensible. Default = `1`}
}
\value{
a plot
}
\description{
`makeWeekdayPlot` returns a plot which shows yVar by weekday, compareYear and plotPeriod (assumed to be lockdown). 
Assumes this is the date aligned data.
}
\seealso{
Other plot: 
\code{\link{addLockdownRect}()},
\code{\link{createDailyMeanComparePlot}()},
\code{\link{createDailyPcComparePlot}()},
\code{\link{createRecentDateTimePlot}()},
\code{\link{createRecentHalfHourlyProfilePlot}()},
\code{\link{makeWeekdayTimePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{plot}
