% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWeekdayTimePlot.R
\name{makeWeekdayTimePlot}
\alias{makeWeekdayTimePlot}
\title{Creates a plot by weekday from the data.table given}
\usage{
makeWeekdayTimePlot(dt, xVar, yVar, yForm = "abs", yLab = "y Lab", yDiv = 1)
}
\arguments{
\item{dt}{the data}

\item{xVar}{the x axis - probably days of the week but you can be creative (needs to be a factor)}

\item{yVar}{the variable you want to plot}

\item{yForm}{whether you want an abs(olute) (default) or a prop(ortional) plot}

\item{yLab}{the label for the y axis}

\item{yDiv}{the value you want to divide yVar by to make the y axis more sensible. Default = `1`}
}
\value{
a plot
}
\description{
`makeWeekdayTimePlot` returns a plot which shows yVar by weekday, time, compareYear and plotPeriod (assumed to be lockdown). 
Assumes this is the date aligned data. Unlike makeWeekdayPlot this plots the proportion of the value (not the absolute value) within
the categories. This helps to see proportional change.
}
\seealso{
Other plot: 
\code{\link{addLockdownRect}()},
\code{\link{createDailyMeanComparePlot}()},
\code{\link{createDailyPcComparePlot}()},
\code{\link{createRecentDateTimePlot}()},
\code{\link{createRecentHalfHourlyProfilePlot}()},
\code{\link{makeWeekdayPlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{plot}
