% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzCalculateCO2e.R
\name{nzCalculateCO2e}
\alias{nzCalculateCO2e}
\title{\code{nzCalculateCO2e} calculates the CO2e emissions per fuel per half-hour.}
\usage{
nzCalculateCO2e(dt, ef)
}
\arguments{
\item{dt}{the data table with the half-hourly kWh values per fuel}

\item{ef}{the MBIE emission factors file}
}
\description{
Uses a long form data.table with each row defined by dateTime & fuel code (and generation station etc).
Assumes kWh, adds variables for carbon intensity (kg CO2e/Kwh) and CO2e (kg) to the data.table and returns it
}
\details{
The calcultion uses imputed conversion factors from the latest NZ efficiency and emission factors published at 
https://www.mbie.govt.nz/building-and-energy/energy-and-natural-resources/energy-statistics-and-modelling/energy-statistics/new-zealand-energy-sector-greenhouse-gas-emissions/

Notes:

 * we estimate the factors each year by dividing the MBIE per-year totals (kt) by the total per-year kWh in the data.table.
If MBIE published the conversion factors this would not be necessary. But if they do, we have not been able to find them.

 * the geothermal conversion does _not_ include fugitive emissions (it probably should)
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getUkGridESO}()},
\code{\link{loadNZEAYearlyGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}

Other utils: 
\code{\link{addLockdownRect}()},
\code{\link{addNZSeason}()},
\code{\link{alignDates}()},
\code{\link{getDuration}()},
\code{\link{gzipFile}()},
\code{\link{loadLibraries}()},
\code{\link{setPeakPeriod}()},
\code{\link{setup}()},
\code{\link{tidyNum}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{data}
\concept{nz}
\concept{utils}
