% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_nzGenMix.R
\name{process_nzGenMix}
\alias{process_nzGenMix}
\title{Load NZ half-hourly generation data by source}
\usage{
process_nzGenMix(dt)
}
\arguments{
\item{dt}{the raw unprocessed data as a data.table}
}
\description{
\code{process_nzGenMix()} processes the wide form files into a long form data.table that is easier to manipulate in R.
}
\details{
We also fix the dateTimes.

Note that rDateTime will be NA for the DST breaks which equate to TP49/50. We really dislike DST breaks.
}
\seealso{
Other data: 
\code{\link{load_nzGenMix}()}

Other grid: 
\code{\link{load_nzGenMix}()}

Other NZ: 
\code{\link{load_nzGenMix}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{NZ}
\concept{data}
\concept{grid}
\concept{utility}
