% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaGen.R
\name{reshapeEAGenDT}
\alias{reshapeEAGenDT}
\title{Reshapes an EA wholesale generation data table from wide (yuk) to long form}
\usage{
reshapeEAGenDT(dt)
}
\description{
\code{reshapeEAGenDT} assumes id.vars exist. They might not
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
