% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaGen.R
\name{reshapeEmbeddedGenDT}
\alias{reshapeEmbeddedGenDT}
\title{Reshapes an EA embedded generation data table from wide (yuk) to long form}
\usage{
reshapeEmbeddedGenDT(dt)
}
\description{
\code{reshapeEmbeddedGenDT} assumes id.vars exist. They might not
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
