% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaGen.R
\name{reshapeGenDT}
\alias{reshapeGenDT}
\title{Reshapes an EA wholesale generation data table from wide (yuk) to long form}
\usage{
reshapeGenDT(dt)
}
\description{
\code{reshapeGenDT} assumes id.vars exist. They might not
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
