% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveUkGridESO.R
\name{saveUkGridESO}
\alias{saveUkGridESO}
\title{\code{saveUkGridESO} saves the UK Grid ESO data to the location you tell it}
\usage{
saveUkGridESO(dt, path)
}
\arguments{
\item{dt}{the data.table to save (returned by \code{cleanUkGridESO})}

\item{path}{the folder to put it in}
}
\description{
\code{saveUkGridESO} saves the UK Grid ESO data to the location you tell it
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getUkGridESO}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{data}
