% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaGen.R
\name{setEAGenTimePeriod}
\alias{setEAGenTimePeriod}
\title{Converts the given time periods (TP1 -> TP48, 49. 50) in a dt to hh:mm}
\usage{
setEAGenTimePeriod(dt)
}
\arguments{
\item{dt}{the data table containing the Time_Period variable}
}
\description{
\code{setEAGenTimePeriod} converts the given time periods (TP1 -> TP48, 49. 50) to hh:mm. It ignores
 TP49 & TP50 as evil incarnations of DST related clock changes (see https://www.emi.ea.govt.nz/Wholesale/Datasets/Generation/Generation_MD/).
 We advise NEVER using the months in which this happens as it will hurt your brain.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
