% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPeakPeriod.R
\name{setPeakPeriod}
\alias{setPeakPeriod}
\title{Codes morning/evening peak periods using rDateTimeUTC in the data.table given}
\usage{
setPeakPeriod(dt, dateTime = rDateTimeUTC)
}
\arguments{
\item{dt}{a data table with a summary variable you want to average}

\item{dateTime}{the variable which is a nice R dateTime for passing to hms::as_hms()}
}
\description{
`setPeakPeriod` adds a factor column `peakPeriod` to the input data.table
}
\seealso{
Other stats: 
\code{\link{create7DayMovingAverage}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{stats}
