% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Set up the project}
\usage{
setup()
}
\description{
\code{setup} sources env.R in the repo/project top level folder.
}
\seealso{
Other utils: 
\code{\link{addLockdownRect}()},
\code{\link{addNZSeason}()},
\code{\link{alignDates}()},
\code{\link{getDuration}()},
\code{\link{gzipFile}()},
\code{\link{loadLibraries}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{setPeakPeriod}()},
\code{\link{tidyNum}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils}
