% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyNum.R
\name{tidyNum}
\alias{tidyNum}
\title{Tidy long numbers}
\usage{
tidyNum(number)
}
\arguments{
\item{number}{an input number or list}
}
\description{
\code{tidyNum} reformats long numbers to include commas and prevents scientific formats
}
\seealso{
Other utils: 
\code{\link{addLockdownRect}()},
\code{\link{addNZSeason}()},
\code{\link{alignDates}()},
\code{\link{getDuration}()},
\code{\link{gzipFile}()},
\code{\link{loadLibraries}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{setPeakPeriod}()},
\code{\link{setup}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils}
