% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLockdownRect.R
\name{addLockdownRect}
\alias{addLockdownRect}
\title{\code{addLockdownRect} adds rectangles around the periods specified as dates or date times on a plot
where the x-axis is a date or datetime. Be sure to specify from & to as dates if the plot x-axis is
a date and as dateTimes if it is a dateTime. Otherwise ggplot will throw a date_trans error.}
\usage{
addLockdownRect(p, from, to, label, yMin, yMax)
}
\arguments{
\item{p}{the plot to add them to}

\item{from}{the start date/datetime}

\item{to}{the end dat dat/datetime}

\item{label}{the label to display}

\item{yMin}{the smallest y value}

\item{yMax}{the largest y value}
}
\description{
\code{addLockdownRect} adds rectangles around the periods specified as dates or date times on a plot
where the x-axis is a date or datetime. Be sure to specify from & to as dates if the plot x-axis is
a date and as dateTimes if it is a dateTime. Otherwise ggplot will throw a date_trans error.
}
\seealso{
Other plot: 
\code{\link{createDailyMeanComparePlot}()},
\code{\link{createDailyPcComparePlot}()},
\code{\link{createRecentDateTimePlot}()},
\code{\link{createRecentHalfHourlyProfilePlot}()},
\code{\link{makeWeekdayPlot}()},
\code{\link{makeWeekdayTimePlot}()}

Other utils: 
\code{\link{addNZSeason}()},
\code{\link{alignDates}()},
\code{\link{getDuration}()},
\code{\link{gzipFile}()},
\code{\link{loadLibraries}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{setPeakPeriod}()},
\code{\link{setup}()},
\code{\link{tidyNum}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{plot}
\concept{utils}
