% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addNZSeason}
\alias{addNZSeason}
\title{Add NZ season to a data.table}
\usage{
addNZSeason(dt, r_dateTime = r_dateTime)
}
\arguments{
\item{dt}{the data table}

\item{r_dateTime}{the column in the dt which is a date that lubridate::month() will work on}
}
\description{
\code{addNZSeason} returns a dt with SOUTHERN hemisphere season added.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
