% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVeDayDateTime.R
\name{addVeDayDateTime}
\alias{addVeDayDateTime}
\title{\code{addVeDayDateTime} adds label for VE day 2020 where x axis is a dateTime. Should work in any time-zone.}
\usage{
addVeDayDateTime(p, yMin, yMax)
}
\arguments{
\item{p}{the plot to add them to}

\item{yMin}{the smallest y value}

\item{yMax}{the largest y value}
}
\description{
\code{addVeDayDateTime} adds label for VE day 2020 where x axis is a dateTime. Should work in any time-zone.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
