% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWeekendRectsDateTime.R
\name{addWeekendRectsDateTime}
\alias{addWeekendRectsDateTime}
\title{\code{addWeekendRectsDateTime} adds rectangles around the weekends from March 2020 onwards
to plots where the x-axis is a dateTime.}
\usage{
addWeekendRectsDateTime(p, yMin, yMax)
}
\arguments{
\item{p}{the plot to add them to}

\item{yMin}{the smallest y value}

\item{yMax}{the largest y value}
}
\description{
\code{addWeekendRectsDateTime} adds rectangles around the weekends from March 2020 onwards
to plots where the x-axis is a dateTime.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
