% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignDates.R
\name{alignDates}
\alias{alignDates}
\title{Re-align 2017-2019 dates so their weekdays match to 2020}
\usage{
alignDates(dt, dateTime)
}
\arguments{
\item{dt}{the data.table to align}

\item{dateTime}{the dateTime variable to use}
}
\description{
\code{alignDates} returns a dt with `fixedDate` set to the date needed in 2017-2019 to match the nearest weekday
in 2020. Use this for plots which try to compare daily values over time.
}
\details{
The data.table to align *must* have a dateTime variable (bu tyou can tell it which one).
}
\seealso{
Other data: 
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getUkGridESO}()},
\code{\link{loadNZEAYearlyEmbeddedGenData}()},
\code{\link{loadNZEAYearlyGridGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}

Other utils: 
\code{\link{addLockdownRect}()},
\code{\link{addNZSeason}()},
\code{\link{getDuration}()},
\code{\link{gzipFile}()},
\code{\link{loadLibraries}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{setPeakPeriod}()},
\code{\link{setup}()},
\code{\link{tidyNum}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{data}
\concept{utils}
