% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUkGridESO.R
\name{cleanUkGridESO}
\alias{cleanUkGridESO}
\title{\code{cleanUkGridESO} cleans up UK Grid ESO data:}
\usage{
cleanUkGridESO(dt)
}
\arguments{
\item{dt}{the data.table to clean (retuirned by \code{getUkGridESO})}
}
\description{
- adds proper dateTime
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
