% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUkGridESO.R
\name{cleanUkGridESO}
\alias{cleanUkGridESO}
\title{\code{cleanUkGridESO} cleans up UK Grid ESO data:}
\usage{
cleanUkGridESO(dt)
}
\arguments{
\item{dt}{the data.table to clean (retuirned by \code{getUkGridESO})}
}
\description{
- adds proper dateTime
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanNZGxpEA}()},
\code{\link{cleanUkEmbeddedESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getNZGxpEA}()},
\code{\link{getNZGxpMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{getUkGridESO}()},
\code{\link{loadEmbeddedGenData}()},
\code{\link{loadNZEAYearlyEmbeddedGenData}()},
\code{\link{loadNZEAYearlyGridGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}

Other grid: 
\code{\link{cleanNZGridEA}()},
\code{\link{cleanNZGxpEA}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getNZGxpEA}()},
\code{\link{getNZGxpMeta}()},
\code{\link{makeNZYearlyData}()},
\code{\link{saveUkGridESO}()}

Other uk: 
\code{\link{cleanUkEmbeddedESO}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{data}
\concept{grid}
\concept{uk}
