% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDailyMeanComparePlot.R
\name{createDailyMeanComparePlot}
\alias{createDailyMeanComparePlot}
\title{Creates a daily half-hourly mean plot from the data.table given}
\usage{
createDailyMeanComparePlot(
  dt,
  yVar,
  yCap,
  form = "step",
  yDiv = 1,
  lockDownStart,
  lockDownEnd,
  comparePlotCut = as.Date("2020-01-01"),
  toDate = lubridate::today()
)
}
\arguments{
\item{dt}{the data, assumed to be the aligned data (use alignDates() to do this)}

\item{yVar}{the variable you want to plot}

\item{yCap}{the caption for the y axis}

\item{form}{do you want a "line" = geom_line() or a "step" = geom_step()? Step is the default}

\item{yDiv}{the value you want to divide yVar by to make the y axis more sensible. Default = 1}

\item{lockDownStart}{date for start of lockdown rectangle annotation}

\item{lockDownEnd}{date for end of lockdown rectangle annotation}

\item{comparePlotCut}{start date to cut the plot (default 2020-01-01)}

\item{toDate}{date to end the plot (usually the most recent observation), default = today()}
}
\description{
`createDailyMeanComparePlot` returns a plot which calculates the mean of yVar and plots it for 2020 and 2017-2019 to 
compare the values for the same date over previous years. This assumes users pass in the aligned-date data (`dateFixed`).
}
\details{
Adds a smoothed line using loess.
}
\seealso{
Other plot: 
\code{\link{addLockdownRect}()},
\code{\link{createDailyPcComparePlot}()},
\code{\link{createRecentDateTimePlot}()},
\code{\link{createRecentHalfHourlyProfilePlot}()},
\code{\link{makeWeekdayPlot}()},
\code{\link{makeWeekdayTimePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{plot}
