% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRecentDateTimePlot.R
\name{createRecentDateTimePlot}
\alias{createRecentDateTimePlot}
\title{Creates a half-hourly plot from the data.table given}
\usage{
createRecentDateTimePlot(
  dt,
  dateTime,
  yVar,
  yCap,
  yDiv = 1,
  lockDownStart,
  lockDownEnd,
  start = as.Date("2020-01-01"),
  end = lubridate::today()
)
}
\arguments{
\item{dt}{the data}

\item{dateTime}{the dateTime variable to use (allows for aligned dateTimes and for other timezones etc etc).}

\item{yVar}{the variable you want to plot}

\item{yCap}{the caption for the y axis}

\item{yDiv}{the value you want to divide yVar by to make the y axis more sensible. Default = `1`}

\item{lockDownStart}{date for start of lockdown rectangle annotation}

\item{lockDownEnd}{date for end of lockdown rectangle annotation}

\item{start}{the start date for the plot}

\item{end}{end date for the plot}
}
\description{
`createRecentDateTimePlot` returns a plot which plots yVar by dateTime and adds lockdown annotations.  This assumes users pass in the aligned-date data (`dateFixed`).
}
\seealso{
Other plot: 
\code{\link{addLockdownRect}()},
\code{\link{createDailyMeanComparePlot}()},
\code{\link{createDailyPcComparePlot}()},
\code{\link{createRecentHalfHourlyProfilePlot}()},
\code{\link{makeWeekdayPlot}()},
\code{\link{makeWeekdayTimePlot}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{plot}
