% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getDuration}
\alias{getDuration}
\title{Get duration}
\usage{
getDuration(t)
}
\arguments{
\item{t}{a duration}
}
\description{
\code{getDuration} takes a timediff created using (e.g.)
}
\details{
startTime <- proc.time()
   .....
   endTime <- proc.time()
   t <- endTime - startTime

and returns a formrated extraction of seconds & minutes for use in feedback.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
