% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNZEmbMeta.R
\name{getNZEmbMeta}
\alias{getNZEmbMeta}
\title{\code{makeNZYearlyData} reports meta data for a given dt}
\usage{
getNZEmbMeta(dt)
}
\arguments{
\item{dt}{data.table to report}
}
\description{
\code{makeNZYearlyData} reports meta data for a given dt
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanNZGxpEA}()},
\code{\link{cleanUkEmbeddedESO}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getNZGxpEA}()},
\code{\link{getNZGxpMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{getUkGridESO}()},
\code{\link{loadEmbeddedGenData}()},
\code{\link{loadNZEAYearlyEmbeddedGenData}()},
\code{\link{loadNZEAYearlyGridGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}

Other embedded: 
\code{\link{cleanNZEmbEA}()},
\code{\link{makeNZYearlyData}()}

Other NZ: 
\code{\link{addNZSeason}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanNZGxpEA}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getNZGxpEA}()},
\code{\link{getNZGxpMeta}()},
\code{\link{makeNZYearlyData}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{NZ}
\concept{data}
\concept{embedded}
