% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNZGridMeta.R
\name{getNZGridMeta}
\alias{getNZGridMeta}
\title{\code{getNZGridMeta} generates metadata for a given grid dt}
\usage{
getNZGridMeta(dt)
}
\arguments{
\item{dt}{the data.table to report}
}
\value{
dt a data.table of info
}
\description{
These are not in pretty form so we clean them up to a long form file and fix the dateTimes forcing to NZT just in case.
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getUkGridESO}()},
\code{\link{loadNZEAYearlyGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}

Other grid: 
\code{\link{cleanNZGridEA}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZGridEA}()},
\code{\link{makeNZYearlyData}()}

Other NZ: 
\code{\link{addNZSeason}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{makeNZYearlyData}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{NZ}
\concept{data}
\concept{grid}
