% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUkGridESO.R
\name{getUkGridESO}
\alias{getUkGridESO}
\title{\code{getUkGridESO} gets the file from the UK Grid ESO data website. These are not in pretty form.
We need to clean them after this.}
\usage{
getUkGridESO(f, update)
}
\arguments{
\item{f}{the file to get (as a url suitable for data.table::fread())}
}
\description{
\code{getUkGridESO} gets the file from the UK Grid ESO data website. These are not in pretty form.
We need to clean them after this.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
