% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzipFile.R
\name{gzipFile}
\alias{gzipFile}
\title{gzip a file}
\usage{
gzipFile(f)
}
\arguments{
\item{f}{the file}
}
\description{
\code{gzipFile} gzips a file
}
\details{
- usually one we've just written out.
}
\seealso{
Other utils: 
\code{\link{addLockdownRect}()},
\code{\link{addNZSeason}()},
\code{\link{alignDates}()},
\code{\link{getDuration}()},
\code{\link{loadLibraries}()},
\code{\link{setPeakPeriod}()},
\code{\link{setup}()},
\code{\link{tidyNum}()}
}
\author{
Michael Harper, \email{m.harper@soton.ac.uk}

Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils}
