% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadUKESOYearlyGenData.R
\name{loadUKESOYearlyGenData}
\alias{loadUKESOYearlyGenData}
\title{Load the pre-downloaded yearly UK ESO generation files in to a data.table}
\usage{
loadUKESOYearlyGenData(path, fromYear, update)
}
\arguments{
\item{path}{the folder to look in for the data}

\item{fromYear}{the year to start from (needs to be in the data file names - you did name them sensibly, yes?)}
}
\description{
\code{loadUKESOYearlyGenData} returns a dt with a proper rDateTimeUTC added & set to London tzone.
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getUkGridESO}()},
\code{\link{loadNZEAYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{data}
