% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadUKESOYearlyGenData.R
\name{loadUKESOYearlyGenData}
\alias{loadUKESOYearlyGenData}
\title{Load the pre-downloaded yearly UK ESO generation files in to a data.table}
\usage{
loadUKESOYearlyGenData(path, fromYear, toDate, update)
}
\arguments{
\item{path}{the folder to look in for the data}

\item{fromYear}{the year to start from (needs to be in the data file names - you did name them sensibly, yes?)}

\item{toDate}{up to but not including this date. e.g. toDate = as.Date("2020-06-01") will be up to and including May 2020}
}
\description{
\code{loadUKESOYearlyGenData} returns a dt with a proper rDateTimeUTC added & set to London tzone.
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanNZGxpEA}()},
\code{\link{cleanUkEmbeddedESO}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getNZGxpEA}()},
\code{\link{getNZGxpMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{getUkGridESO}()},
\code{\link{loadEmbeddedGenData}()},
\code{\link{loadNZEAYearlyEmbeddedGenData}()},
\code{\link{loadNZEAYearlyGridGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{data}
