% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNZYearlyData.R
\name{makeNZYearlyData}
\alias{makeNZYearlyData}
\title{\code{makeNZYearlyData} converts monthly to yearly data}
\usage{
makeNZYearlyData(genType, years, path)
}
\arguments{
\item{genType}{understands `gridGen` or `embeddedGen`}

\item{years}{which years to loop over}

\item{path}{the top level data path for this genType}
}
\description{
\code{makeNZYearlyData} converts monthly to yearly data
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getUkGridESO}()},
\code{\link{loadNZEAYearlyGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}

Other embedded: 
\code{\link{cleanNZEmbEA}()},
\code{\link{getNZEmbMeta}()}

Other grid: 
\code{\link{cleanNZGridEA}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()}

Other NZ: 
\code{\link{addNZSeason}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{NZ}
\concept{data}
\concept{embedded}
\concept{grid}
