% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUkGridESOYearlyData.R
\name{makeUkGridESOYearlyData}
\alias{makeUkGridESOYearlyData}
\title{\code{makeUkGridESOYearlyData} saves the UK Grid ESO data as yearly files to the location you tell it}
\usage{
makeUkGridESOYearlyData(dt, path)
}
\arguments{
\item{dt}{the data.table to save (returned by \code{cleanUkGridESO})}

\item{path}{the folder to put the yearly files in}
}
\description{
\code{makeUkGridESOYearlyData} saves the UK Grid ESO data as yearly files to the location you tell it
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getUkGridESO}()},
\code{\link{loadNZEAYearlyGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{data}
