% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzCalculateCO2e.R
\name{nzCalculateCO2e}
\alias{nzCalculateCO2e}
\title{\code{nzCalculateCO2e} calculates the CO2e emissions (in kg) per fuel per half-hour.}
\usage{
nzCalculateCO2e(dt, ef)
}
\arguments{
\item{dt}{the data table with the half-hourly kWh values per fuel (do not include incomplete years)}

\item{ef}{the MBIE emission factors file}
}
\description{
Uses the EA grid generation data to sum yearly kWh by fuel and combines with the NZ efficiency and emission factors published at 
https://www.mbie.govt.nz/building-and-energy/energy-and-natural-resources/energy-statistics-and-modelling/energy-statistics/new-zealand-energy-sector-greenhouse-gas-emissions/
}
\details{
Returns a data.table of carbon intensity (gCO2-e/kWh) per fuel per year for further use
 
Notes:

 * we estimate the factors each year by dividing the MBIE per-year totals (kt CO2-e) by the total per fuel per-year kWh in the data.table and then converting
 to g/kWh. If MBIE published the conversion factors this would not be necessary. But if they do, we have not been able to find them.

 * the geothermal conversion _may not_ include fugitive emissions (?)
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanUkEmbeddedESO}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getUkEmbeddedESO}()},
\code{\link{getUkGridESO}()},
\code{\link{loadEmbeddedGenData}()},
\code{\link{loadNZEAYearlyEmbeddedGenData}()},
\code{\link{loadNZEAYearlyGridGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{reshapeEmbeddedGenDT}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkEmbeddedESO}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}

Other utils: 
\code{\link{addLockdownRect}()},
\code{\link{addNZSeason}()},
\code{\link{addSeason}()},
\code{\link{alignDates}()},
\code{\link{getDuration}()},
\code{\link{gzipFile}()},
\code{\link{loadLibraries}()},
\code{\link{setPeakPeriod}()},
\code{\link{setup}()},
\code{\link{tidyNum}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{data}
\concept{nz}
\concept{utils}
