% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshapeEmbeddedGenDT.R
\name{reshapeEmbeddedGenDT}
\alias{reshapeEmbeddedGenDT}
\title{Reshapes an EA embedded generation data table from wide (yuk) to long form}
\usage{
reshapeEmbeddedGenDT(dt)
}
\description{
\code{reshapeEmbeddedGenDT} assumes id.vars exist. They might not
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getUkGridESO}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{data}
