% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshapeEmbeddedGenDT.R
\name{reshapeEmbeddedGenDT}
\alias{reshapeEmbeddedGenDT}
\title{Reshapes an EA embedded generation data table from wide (yuk) to long form}
\usage{
reshapeEmbeddedGenDT(dt)
}
\description{
\code{reshapeEmbeddedGenDT} assumes id.vars exist. They might not
}
\seealso{
Other data: 
\code{\link{alignDates}()},
\code{\link{cleanNZEmbEA}()},
\code{\link{cleanNZGridEA}()},
\code{\link{cleanUkGridESO}()},
\code{\link{getNZEmbData}()},
\code{\link{getNZEmbMeta}()},
\code{\link{getNZGridEA}()},
\code{\link{getNZGridMeta}()},
\code{\link{getUkGridESO}()},
\code{\link{loadNZEAYearlyGenData}()},
\code{\link{loadUKESOYearlyGenData}()},
\code{\link{makeNZYearlyData}()},
\code{\link{makeUkGridESOYearlyData}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{reshapeGenDT}()},
\code{\link{saveUkGridESO}()},
\code{\link{setEmbeddedGenTimePeriod}()},
\code{\link{setGridGenTimePeriod}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
\concept{data}
