% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPeakPeriod.R
\name{setPeakPeriod}
\alias{setPeakPeriod}
\title{Codes morning/evening peak periods using rDateTimeUTC in the data.table given}
\usage{
setPeakPeriod(
  dt,
  dateTime = rDateTimeUTC,
  t1 = "07:00",
  t2 = "09:00",
  t3 = "16:00",
  t4 = "20:00"
)
}
\arguments{
\item{dt}{a data table with a summary variable you want to average}

\item{dateTime}{the variable which is a nice R dateTime for passing to hms::as_hms()}

\item{t1}{start of morning peak in 24 hour clock e.g. "07:00" (default)}

\item{t2}{end of morning peak in 24 hour clock e.g. "08:30" (default)}

\item{t3}{start of morning peak in 24 hour clock e.g. "16:00" (default)}

\item{t4}{start of morning peak in 24 hour clock e.g. "20:00" (default)}
}
\description{
`setPeakPeriod` adds a factor column `peakPeriod` to the input data.table
}
\seealso{
Other utils: 
\code{\link{addLockdownRect}()},
\code{\link{addNZSeason}()},
\code{\link{addSeason}()},
\code{\link{alignDates}()},
\code{\link{getDuration}()},
\code{\link{gzipFile}()},
\code{\link{loadLibraries}()},
\code{\link{nzCalculateCO2e}()},
\code{\link{setup}()},
\code{\link{tidyNum}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils}
