% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Set up the project}
\usage{
setup()
}
\description{
\code{setup} sources env.R in the repo/project top level folder.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{utils}
